Chapter 1, Task 1 - Implement the get_movies_by_country method in db.py to search movies by country and use projection to return the title field. Chapter 1, Task 2 - For this ticket, you will need to modify the method build_query_sort_project in db.py to allow the following movie search criteria:
genres: finds all movies that match the search genres
Already, the build_query_sort_project method is able to return results for two different types of movie search criteria:

text : performs a text search in the movies collection
cast: finds movies that include any of the wanted cast
You just need to construct the query that queries the movies collection by genre.

A text index was created for you when you restored the collections with mongorestore, so these queries will be performant once they are implemented.